// src/api/api.ts
export interface CategoryResponseDTO {
    'id'?: string;
    'name'?: string;
    'type'?: CategoryResponseDTOTypeEnum;
    'color'?: string;
}

export const CategoryResponseDTOTypeEnum = {
    Despesa: 'DESPESA',
    Receita: 'RECEITA'
} as const;

export type CategoryResponseDTOTypeEnum = typeof CategoryResponseDTOTypeEnum[keyof typeof CategoryResponseDTOTypeEnum];

export interface CategoryTotalDTO {
    'name'?: string;
    'total'?: number;
    'type'?: CategoryTotalDTOTypeEnum;
}

export const CategoryTotalDTOTypeEnum = {
    Despesa: 'DESPESA',
    Receita: 'RECEITA'
} as const;

export type CategoryTotalDTOTypeEnum = typeof CategoryTotalDTOTypeEnum[keyof typeof CategoryTotalDTOTypeEnum];

export interface CreateCategoryDTO {
    'name': string;
    'type': CreateCategoryDTOTypeEnum;
    'color'?: string;
}

export const CreateCategoryDTOTypeEnum = {
    Despesa: 'DESPESA',
    Receita: 'RECEITA'
} as const;

export type CreateCategoryDTOTypeEnum = typeof CreateCategoryDTOTypeEnum[keyof typeof CreateCategoryDTOTypeEnum];

export interface CreateTransactionDTO {
    'description': string;
    'date': string;
    'amount': number;
    'category_id': string;
}
// ...
export interface Pageable {
    'page'?: number;
    'size'?: number;
    'sort'?: Array<string>;
}
// ...
export interface UpdateAmountTransactionDTO {
    'amount': number;
}
export interface UpdateCategoryDTO {
    'name': string;
    'type': UpdateCategoryDTOTypeEnum;
    'color'?: string;
}

export const UpdateCategoryDTOTypeEnum = {
    Despesa: 'DESPESA',
    Receita: 'RECEITA'
} as const;

export type UpdateCategoryDTOTypeEnum = typeof UpdateCategoryDTOTypeEnum[keyof typeof UpdateCategoryDTOTypeEnum];

export interface UpdateDateTransactionDTO {
    'date': string;
}
export interface UpdateDescriptionTransactionDTO {
    'description': string;
}
export interface UpdateEmailAndUsernameDTO {
    'username': string;
    'email': string;
}
export interface UpdateEmailDTO {
    'email': string;
}
export interface UpdateNameCategoryDTO {
    'name': string;
}
export interface UpdatePasswordDTO {
    'currentPassword': string;
    'password': string;
    'confirmPassword': string;
}
export interface UpdateTransactionDTO {
    'description': string;
    'date': string;
    'amount': number;
    'category_id': string;
}
export interface UpdateTypeCategoryDTO {
    'type': UpdateTypeCategoryDTOTypeEnum;
}

export const UpdateTypeCategoryDTOTypeEnum = {
    Despesa: 'DESPESA',
    Receita: 'RECEITA'
} as const;

export type UpdateTypeCategoryDTOTypeEnum = typeof UpdateTypeCategoryDTOTypeEnum[keyof typeof UpdateTypeCategoryDTOTypeEnum];
// ...
export class CategoriaApi extends BaseAPI {
    /**
     * 
     * @param {CreateCategoryDTO} createCategoryDTO 
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     */
    public createCategory(createCategoryDTO: CreateCategoryDTO, options?: RawAxiosRequestConfig) {
        return CategoriaApiFp(this.configuration).createCategory(createCategoryDTO, options).then((request) => request(this.axios, this.basePath));
    }

    /**
     * 
     * @param {string} categoryId 
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     */
    public deleteCategory(categoryId: string, options?: RawAxiosRequestConfig) {
        return CategoriaApiFp(this.configuration).deleteCategory(categoryId, options).then((request) => request(this.axios, this.basePath));
    }

    /**
     * 
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     */
    public getAllCategories(options?: RawAxiosRequestConfig) {
        return CategoriaApiFp(this.configuration).getAllCategories(options).then((request) => request(this.axios, this.basePath));
    }

    /**
     * 
     * @param {string} categoryId 
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     */
    public getCategoryById(categoryId: string, options?: RawAxiosRequestConfig) {
        return CategoriaApiFp(this.configuration).getCategoryById(categoryId, options).then((request) => request(this.axios, this.basePath));
    }

    /**
     * 
     * @param {string} categoryId 
     * @param {UpdateCategoryDTO} updateCategoryDTO 
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     */
    public updateCategory(categoryId: string, updateCategoryDTO: UpdateCategoryDTO, options?: RawAxiosRequestConfig) {
        return CategoriaApiFp(this.configuration).updateCategory(categoryId, updateCategoryDTO, options).then((request) => request(this.axios, this.basePath));
    }

    /**
     * 
     * @param {string} categoryId 
     * @param {UpdateNameCategoryDTO} updateNameCategoryDTO 
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     */
    public updateCategoryName(categoryId: string, updateNameCategoryDTO: UpdateNameCategoryDTO, options?: RawAxiosRequestConfig) {
        return CategoriaApiFp(this.configuration).updateCategoryName(categoryId, updateNameCategoryDTO, options).then((request) => request(this.axios, this.basePath));
    }

    /**
     * 
     * @param {string} categoryId 
     * @param {UpdateTypeCategoryDTO} updateTypeCategoryDTO 
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     */
    public updateCategoryType(categoryId: string, updateTypeCategoryDTO: UpdateTypeCategoryDTO, options?: RawAxiosRequestConfig) {
        return CategoriaApiFp(this.configuration).updateCategoryType(categoryId, updateTypeCategoryDTO, options).then((request) => request(this.axios, this.basePath));
    }
}
// ...
export class TransacaoApi extends BaseAPI {
    /**
     * 
     * @param {string} transactionId 
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     */
    public _delete(transactionId: string, options?: RawAxiosRequestConfig) {
        return TransacaoApiFp(this.configuration)._delete(transactionId, options).then((request) => request(this.axios, this.basePath));
    }

    /**
     * 
     * @param {CreateTransactionDTO} createTransactionDTO 
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     */
    public create(createTransactionDTO: CreateTransactionDTO, options?: RawAxiosRequestConfig) {
        return TransacaoApiFp(this.configuration).create(createTransactionDTO, options).then((request) => request(this.axios, this.basePath));
    }

    /**
     * 
     * @param {Pageable} pageable 
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     */
    public getAllPageSortingByDate(pageable: Pageable, options?: RawAxiosRequestConfig) {
        return TransacaoApiFp(this.configuration).getAllPageSortingByDate(pageable, options).then((request) => request(this.axios, this.basePath));
    }

    /**
     * 
     * @param {string} transactionId 
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     */
    public getById(transactionId: string, options?: RawAxiosRequestConfig) {
        return TransacaoApiFp(this.configuration).getById(transactionId, options).then((request) => request(this.axios, this.basePath));
    }

    /**
     * 
     * @param {File} file 
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     */
    public importXlsx(file: File, options?: RawAxiosRequestConfig) {
        return TransacaoApiFp(this.configuration).importXlsx(file, options).then((request) => request(this.axios, this.basePath));
    }

    /**
     * 
     * @param {string} transactionId 
     * @param {UpdateAmountTransactionDTO} updateAmountTransactionDTO 
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     */
    public updateAmount(transactionId: string, updateAmountTransactionDTO: UpdateAmountTransactionDTO, options?: RawAxiosRequestConfig) {
        return TransacaoApiFp(this.configuration).updateAmount(transactionId, updateAmountTransactionDTO, options).then((request) => request(this.axios, this.basePath));
    }

    /**
     * 
     * @param {string} transactionId 
     * @param {UpdateDateTransactionDTO} updateDateTransactionDTO 
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     */
    public updateDate(transactionId: string, updateDateTransactionDTO: UpdateDateTransactionDTO, options?: RawAxiosRequestConfig) {
        return TransacaoApiFp(this.configuration).updateDate(transactionId, updateDateTransactionDTO, options).then((request) => request(this.axios, this.basePath));
    }

    /**
     * 
     * @param {string} transactionId 
     * @param {UpdateDescriptionTransactionDTO} updateDescriptionTransactionDTO 
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     */
    public updateDescription(transactionId: string, updateDescriptionTransactionDTO: UpdateDescriptionTransactionDTO, options?: RawAxiosRequestConfig) {
        return TransacaoApiFp(this.configuration).updateDescription(transactionId, updateDescriptionTransactionDTO, options).then((request) => request(this.axios, this.basePath));
    }

    /**
     * 
     * @param {string} transactionId 
     * @param {UpdateTransactionDTO} updateTransactionDTO 
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     */
    public updateTransaction(transactionId: string, updateTransactionDTO: UpdateTransactionDTO, options?: RawAxiosRequestConfig) {
        return TransacaoApiFp(this.configuration).updateTransaction(transactionId, updateTransactionDTO, options).then((request) => request(this.axios, this.basePath));
    }
}